(piqi) Serialize .piqi as piqi-spec instead of piqi-lang
When serializing .piqi to JSON, XML or PB, automatically expand it by resolving
includes and extensions.

By doing this, we ensure that the full .piqi specification will be available to
those who want to work with it using languages other than OCaml.

Resolving extensions and includes is necessary, because there is no convenient
way to represent Piqi extensions in those formats (serialized Piq ast doesn't
make a lot of sense to users). In addition, not having to deal with includes
makes working with Piqi specs really easy and straighforward.